Added unit test project with one unti test.  Had to refactor controller so that code was testable.
