+++ WebService.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29519.87
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebService", "WebService\WebService.csproj", "{F848E04F-6731-46E2-BF65-1C4C1E98DD27}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebService", "WebService\WebService.csproj", "{F848E04F-6731-46E2-BF65-1C4C1E98DD27}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{6F3722A3-3C22-40E8-A5C9-581C552CE197}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{F848E04F-6731-46E2-BF65-1C4C1E98DD27}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F848E04F-6731-46E2-BF65-1C4C1E98DD27}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F848E04F-6731-46E2-BF65-1C4C1E98DD27}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F3722A3-3C22-40E8-A5C9-581C552CE197}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F3722A3-3C22-40E8-A5C9-581C552CE197}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F3722A3-3C22-40E8-A5C9-581C552CE197}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F3722A3-3C22-40E8-A5C9-581C552CE197}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
