+++ WebService/Controllers/WeatherForecastController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using WebService.Models;
 
 namespace Project.WebService.Controllers
 {
@@ -11,10 +12,10 @@ namespace Project.WebService.Controllers
     [Route("[controller]")]
     public class WeatherForecastController : ControllerBase
     {
-        private static readonly string[] Summaries = new[]
-        {
-            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-        };
+        //private static readonly string[] Summaries = new[]
+        //{
+        //    "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        //};
 
         private readonly ILogger<WeatherForecastController> _logger;
 
@@ -27,17 +28,19 @@ namespace Project.WebService.Controllers
         public IEnumerable<WeatherForecast> Get()
         {
             _logger.LogInformation($"Begin GET/weatherforecast @ {DateTime.Now}");
-            var rng = new Random();
-            var weather = Enumerable.Range(1, 5).Select(index => new WeatherForecast
-            {
-                Date = DateTime.Now.AddDays(index),
-                TemperatureC = rng.Next(-20, 55),
-                Summary = Summaries[rng.Next(Summaries.Length)]
-            })
-            .ToArray();
+            //var rng = new Random();
+            //var weather = Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            //{
+            //    Date = DateTime.Now.AddDays(index),
+            //    TemperatureC = rng.Next(-20, 55),
+            //    Summary = Summaries[rng.Next(Summaries.Length)]
+            //})
+            //.ToArray();
+            var forecaster = new WeatherForecaster();
+
             _logger.LogInformation($"End GET/weatherforecast @ {DateTime.Now}");
 
-            return weather;
+            return forecaster.WeatherSummaires();
         }
     }
 }
